pyproject.toml
[project]
name = "mini-rag-backend"
version = "0.1.0"
description = "RAG backend (FastAPI + LlamaIndex + Supabase) with HTTP-only model calls."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "llama-index-core>=0.11",
  "llama-index-vector-stores-supabase>=0.2",
  "llama-index-llms-openai>=0.2",
  "llama-index-embeddings-openai>=0.2",
  "supabase>=2.6",
  "python-dotenv>=1.0",
  "pymupdf>=1.24",
  "pdfplumber>=0.11",
  "beautifulsoup4>=4.12",
  "html2text>=2024.2",
  "python-docx>=1.1",
  "openpyxl>=3.1",
  "pyyaml>=6.0",
  "python-multipart>=0.0.26",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.ruff.lint]
select = ["E", "F", "I"]

